ARM NEON optimization for float->int conversion.
Not optimal, but since lrintf is incredibly slow on
ARM it is > 10x faster than the old code.
A fallback solution that (incorrectly) defines lrintf(x)
as (int)(x) might make sense to avoid these kind of issues
for the pure C code, however we would still need to know whether
lrintf is slow or not.
Though maybe the better solution was if all architectures provided
a non-braindead implementation of lrintf.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@36489 b3059339-0415-0410-9bf9-f77b7e298cf2