Fix the index passed to PutImage() for item type itVPotmeter.
Unlike the position for drawing the button (from bottom to top), the
index for the phase image has to be in the usual range, i.e. 0 for 0%
and numphases - 1 for 100%, so that the phase image can be designed
straightforwardly.

Fix the parameter although PutImage() currently doesn't work for item
type itVPotmeter at all.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37099 b3059339-0415-0410-9bf9-f77b7e298cf2