demux_real: Improve buffer allocation for interleaved audio handling.
Allocate buffers where they are needed.
This avoids code duplication, allocating them when they are not
needed, and might avoid a crash if an audio stream is specified
via -aid that does not exist in the headers but does exist in the
file since then we might run the deinterleaving without having
the buffers allocated first (note: I have not tested this can
actually happen).

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37200 b3059339-0415-0410-9bf9-f77b7e298cf2