Fix playback of hls stream where url is longer than 256.
MPlayer was not able to update the m3u8 playlist when the lenght of the url is
greater than 256, because mp_filename array was only 256 bytes.
The patch simply increase the size of mp_filename to 2048, the value matches
the size used in ffmpeg url.

Patch by Lu Ran <hephooey gmail>



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37870 b3059339-0415-0410-9bf9-f77b7e298cf2